gotox/secapi: Fix rotor control
I'm testing the dvbsec api for the Goto X part to control my rotor
and found it did not respond as it should be.

I pulled the DiseqC 1.2 spec from the Eutelsat website and found a
bug in the dvbsec_api.c file. The first diff file solves the bug
and makes the api follow the spec as it should be. But after testing
I found out my rotor did not what it should do.

From: Michel Verbraak <michel@verbraak.org>