libdvben50221: Correctly free resources when a session is closed. This allows a CA module to re-open a session.
The issue was, that libdvben50221 did not allow a CAM to re-establish
the session holding non-MMI resources if using the low level interface.

The session number was recorded on open, but not freed on close
(which IMO) is a bug in the code. With the following change, a SMIT CAM
with a Showtime card works fine according to tests so far.

The effect was, that the CAM tried to constantly close and re-open the
session and libdvben50221 kept telling it, that the resource is already
allocated to a different session. Additionally this caused the library
to use the OLD session number in communications with the CAM, which did
not even exist anymore, so caused all writes of CA PMT objects to fail
with -EINTR.

From: Stephan Trebels <stephan <at> trebels.com>